requirements.txt
aws-cdk-lib
constructs>=10.0.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v27
